Carly's Clippers
You are the Data Analyst at Carly’s Clippers, the newest hair salon on the block. Your job is to go through the lists of data that have been collected in the past couple of weeks. You will be calculating some important metrics that Carly can use to plan out the operation of the business for the rest of the month.

You have been provided with three lists:

    hairstyles: the names of the cuts offered at Carly’s Clippers.
    prices: the price of each hairstyle in the hairstyles list.
    last_week: the number of purchases for each hairstyle type in the last week.

Each index in hairstyles corresponds to an associated index in prices and last_week.

For example, The hairstyle "bouffant" has an associated price of 30 from the prices list, and was purchased 2 times in the last week as shown in the last_week list. Each of these elements are in the first index of their respective lists.

-Carly wants to be able to market her low prices. We want to find out what the average price of a cut is.
-That average price is more expensive than Carly thought it would be! She wants to cut all prices by 5 dollars.
-Carly really wants to make sure that Carly’s Clippers is a profitable endeavor. She first wants to know how much revenue was brought in last week.
-Carly thinks she can bring in more customers by advertising all of the haircuts she has that are under 30 dollars.